+++ crypto/src/commitments/kzg.rs
@@ -1,139 +1,13 @@
-use crate::errors::SrsFromFileError;
 
-use super::traits::IsCommitmentScheme;
+use super::{traits::IsCommitmentScheme, srs::StructuredReferenceString};
 use lambdaworks_math::{
     cyclic_group::IsGroup,
     elliptic_curve::traits::IsPairing,
-    errors::DeserializationError,
     field::{element::FieldElement, traits::IsPrimeField},
     msm::pippenger::msm,
-    polynomial::Polynomial,
-    traits::{Deserializable, Serializable},
-    unsigned_integer::element::UnsignedInteger,
+    unsigned_integer::element::UnsignedInteger, polynomial::Polynomial,
 };
-use std::{marker::PhantomData, mem};
-
-#[derive(PartialEq, Clone, Debug)]
-pub struct StructuredReferenceString<G1Point, G2Point> {
-    pub powers_main_group: Vec<G1Point>,
-    pub powers_secondary_group: [G2Point; 2],
-}
-
-impl<G1Point, G2Point> StructuredReferenceString<G1Point, G2Point>
-where
-    G1Point: IsGroup,
-    G2Point: IsGroup,
-{
-    pub fn new(powers_main_group: &[G1Point], powers_secondary_group: &[G2Point; 2]) -> Self {
-        Self {
-            powers_main_group: powers_main_group.into(),
-            powers_secondary_group: powers_secondary_group.clone(),
-        }
-    }
-}
-
-impl<G1Point, G2Point> StructuredReferenceString<G1Point, G2Point>
-where
-    G1Point: IsGroup + Deserializable,
-    G2Point: IsGroup + Deserializable,
-{
-    pub fn from_file(file_path: &str) -> Result<Self, SrsFromFileError> {
-        let bytes = std::fs::read(file_path)?;
-        Ok(Self::deserialize(&bytes)?)
-    }
-}
-
-impl<G1Point, G2Point> Serializable for StructuredReferenceString<G1Point, G2Point>
-where
-    G1Point: IsGroup + Serializable,
-    G2Point: IsGroup + Serializable,
-{
-    fn serialize(&self) -> Vec<u8> {
-        let mut serialized_data: Vec<u8> = Vec::new();
-        // First 4 bytes encodes protocol version
-        let protocol_version: [u8; 4] = [0; 4];
-
-        serialized_data.extend(&protocol_version);
-
-        // Second 8 bytes store the amount of G1 elements to be stored, this is more than can be indexed with a 64-bit architecture, and some millions of terabytes of data if the points were compressed
-        let mut main_group_len_bytes: Vec<u8> = self.powers_main_group.len().to_le_bytes().to_vec();
-
-        // For architectures with less than 64 bits for pointers
-        // We add extra zeros at the end
-        while main_group_len_bytes.len() < 8 {
-            main_group_len_bytes.push(0)
-        }
-
-        serialized_data.extend(&main_group_len_bytes);
-
-        // G1 elements
-        for point in &self.powers_main_group {
-            serialized_data.extend(point.serialize());
-        }
-
-        // G2 elements
-        for point in &self.powers_secondary_group {
-            serialized_data.extend(point.serialize());
-        }
-
-        serialized_data
-    }
-}
-
-impl<G1Point, G2Point> Deserializable for StructuredReferenceString<G1Point, G2Point>
-where
-    G1Point: IsGroup + Deserializable,
-    G2Point: IsGroup + Deserializable,
-{
-    fn deserialize(bytes: &[u8]) -> Result<Self, DeserializationError> {
-        const MAIN_GROUP_LEN_OFFSET: usize = 4;
-        const MAIN_GROUP_OFFSET: usize = 12;
-
-        let main_group_len_u64 = u64::from_le_bytes(
-            // This unwrap can't fail since we are fixing the size of the slice
-            bytes[MAIN_GROUP_LEN_OFFSET..MAIN_GROUP_OFFSET]
-                .try_into()
-                .unwrap(),
-        );
-
-        let main_group_len = usize::try_from(main_group_len_u64)
-            .map_err(|_| DeserializationError::PointerSizeError)?;
-
-        let mut main_group: Vec<G1Point> = Vec::new();
-        let mut secondary_group: Vec<G2Point> = Vec::new();
-
-        let size_g1_point = mem::size_of::<G1Point>();
-        let size_g2_point = mem::size_of::<G2Point>();
-
-        for i in 0..main_group_len {
-            // The second unwrap shouldn't fail since the amount of bytes is fixed
-            let point = G1Point::deserialize(
-                bytes[i * size_g1_point + MAIN_GROUP_OFFSET
-                    ..i * size_g1_point + size_g1_point + MAIN_GROUP_OFFSET]
-                    .try_into()
-                    .unwrap(),
-            )?;
-            main_group.push(point);
-        }
-
-        let g2s_offset = size_g1_point * main_group_len + 12;
-        for i in 0..2 {
-            // The second unwrap shouldn't fail since the amount of bytes is fixed
-            let point = G2Point::deserialize(
-                bytes[i * size_g2_point + g2s_offset
-                    ..i * size_g2_point + g2s_offset + size_g2_point]
-                    .try_into()
-                    .unwrap(),
-            )?;
-            secondary_group.push(point);
-        }
-
-        let secondary_group_slice = [secondary_group[0].clone(), secondary_group[1].clone()];
-
-        let srs = StructuredReferenceString::new(&main_group, &secondary_group_slice);
-        Ok(srs)
-    }
-}
+use std::marker::PhantomData;
 
 #[derive(Clone)]
 pub struct KateZaveruchaGoldberg<F: IsPrimeField, P: IsPairing> {
