+++ crypto/src/commitments/mod.rs
@@ -1,2 +1,3 @@
 pub mod kzg;
 pub mod traits;
+pub mod srs;
