+++ crypto/src/lookups/cq/mod.rs
@@ -1,5 +1,7 @@
 pub mod setup;
+pub mod gen;
 pub mod prover;
 pub mod verifier;
 pub mod data_structures;
 pub mod errors;
+pub mod utils;
